[ROCm/MIOpen commit: 162c67993774a5ba1f438387b9811abf712b341d]

+++ src/include/miopen/tensor.hpp
@@ -78,7 +78,7 @@ struct TensorDescriptor : miopenTensorDescriptor
     template <class... Ts>
     std::size_t GetIndex(Ts... is) const
     {
-        return this->GetIndex({is...});
+        return this->GetIndex({static_cast<int>(is)...});
     }
 
     bool operator==(const TensorDescriptor& rhs) const;
